backend/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
python-dotenv==1.0.0
SQLAlchemy==2.0.20
aiosqlite==0.18.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0
requests==2.31.0
openai==1.0.0
numpy==1.26.0
